TST: run: Move tests that don't need rerun to core.local
The run module is now in core.local.  We can't move all the run tests
to core.local because many of them also test rerun, but move the ones
that don't test rerun.

As shown by

    git show --no-renames --color-moved ...

there are no changes to any tests.
